move torrentCheck into its own file, refactor trackerOpts
